Reverse payload-order to token-timestamp-useridentifier
Previous to this commit the payload was constructed as a userprovided
identifier, a unique token and a timestamp, concatenated by - in this
order. As the payload was started with the user-provided identifier,
parsing the payload was requiring said identifier not to contain the
delimiter, as otherwise parsing would break. By switching the order,
the mailexporter contains the first two parts of the message and we can
safely split twice and treat the complete third part of the payload as
the user-provided identifier, regardless of characters included.
Fixes #29.